[refactor] Moved 'removeChildrenFromElements' to its own TS file
