+++ src/CompanyManagement.js
@@ -7,8 +7,7 @@ import {Player}                                         from "./Player";
 
 import Decimal                                          from "decimal.js";
 import {dialogBoxCreate}                                from "../utils/DialogBox";
-import {removeChildrenFromElement,
-        clearSelector}                                  from "../utils/HelperFunctions";
+import {clearSelector}                                  from "../utils/HelperFunctions";
 import {Reviver, Generic_toJSON,
         Generic_fromJSON}                               from "../utils/JSONReviver";
 import numeral                                          from "numeral/min/numeral.min";
@@ -17,6 +16,7 @@ import {createPopup}                                    from "../utils/uiHelpers
 import {formatNumber, generateRandomString}             from "../utils/StringHelperFunctions";
 import {getRandomInt}                                   from "../utils/helpers/getRandomInt";
 import {isString}                                       from "../utils/helpers/isString";
+import {removeChildrenFromElement}                      from "../utils/uiHelpers/removeChildrenFromElement";
 import {removeElementById}                              from "../utils/uiHelpers/removeElementById";
 import {yesNoBoxCreate, yesNoTxtInpBoxCreate,
         yesNoBoxGetYesButton, yesNoBoxGetNoButton,
